Extract review decision module
